update favorites function

+++ src/Context/Context.jsx
@@ -59,27 +59,7 @@ const Context = ({ children }) => {
   };
 
   const addToFavorites = async (item) => {
-    try {
-      await axios.post(
-        "https://matiastmbdback.onrender.com/addFavorites",
-        {
-          email: userLogged.data.email,
-          movieId: item.id,
-          title: item.title || item.name,
-          vote_average: item.vote_average,
-          poster_path: item.poster_path,
-          adult: item.title ? true : false,
-        },
-        { withCredentials: true, credentials: "include" }
-      );
-      toast.success("Successfully added to favorites", {
-        duration: "100",
-        style: {
-          background: "black",
-          color: "white",
-        },
-      });
-    } catch {
+    if (userLogged !== {}) {
       toast.error("You must log in to add favorites", {
         duration: "100",
         style: {
@@ -87,22 +67,69 @@ const Context = ({ children }) => {
           color: "white",
         },
       });
+      return;
+    }
+    try {
+      const isFavMovie = userFavorites.find(
+        (userMovie) => userMovie.id === item.id
+      );
+      if (isFavMovie) {
+        const newFavorites = userFavorites.filter(
+          (userMovie) => userMovie.id !== item.id
+        );
+        setUserFavorites(newFavorites);
+        await axios.delete(
+          `https://matiastmbdback.onrender.com/removeFavorites?id=${item.id}`,
+          { withCredentials: true, credentials: "include" }
+        );
+        toast.error("Successfully deleted from favorites", {
+          duration: "100",
+          style: {
+            background: "black",
+            color: "white",
+          },
+        });
+      } else {
+        const newFavorites = [...userFavorites, item];
+        setUserFavorites(newFavorites);
+        await axios.post(
+          "https://matiastmbdback.onrender.com/addFavorites",
+          {
+            email: userLogged.data.email,
+            movieId: item.id,
+            title: item.title || item.name,
+            vote_average: item.vote_average,
+            poster_path: item.poster_path,
+            adult: item.title ? true : false,
+          },
+          { withCredentials: true, credentials: "include" }
+        );
+        toast.success("Successfully added to favorites", {
+          duration: "100",
+          style: {
+            background: "black",
+            color: "white",
+          },
+        });
+      }
+    } catch (error) {
+      console.error(error);
     }
   };
 
-  const removeFromFavorites = async (item) => {
-    await axios.delete(
-      `https://matiastmbdback.onrender.com/removeFavorites?id=${item.id}`,
-      { withCredentials: true, credentials: "include" }
-    );
-    toast.error("Successfully deleted from favorites", {
-      duration: "100",
-      style: {
-        background: "black",
-        color: "white",
-      },
-    });
-  };
+  // const removeFromFavorites = async (item) => {
+  //   await axios.delete(
+  //     `https://matiastmbdback.onrender.com/removeFavorites?id=${item.id}`,
+  //     { withCredentials: true, credentials: "include" }
+  //   );
+  //   toast.error("Successfully deleted from favorites", {
+  //     duration: "100",
+  //     style: {
+  //       background: "black",
+  //       color: "white",
+  //     },
+  //   });
+  // };
 
   const valueContext = {
     userLogged,
@@ -116,7 +143,7 @@ const Context = ({ children }) => {
     userFavorites,
     setUserFavorites,
     addToFavorites,
-    removeFromFavorites,
+    // removeFromFavorites,
   };
   return <Provider value={valueContext}>{children}</Provider>;
 };
